# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FollowUpBussiness API
  version: 1.0.0
  summary: Contrato REST general del MVP
  description: |
    API multiempresa para administración, aplicación móvil y procesos de
    integración de FollowUpBussiness.

    Reglas globales:
    - El tenant y el usuario se derivan de la identidad autenticada.
    - Ningún `tenantId` recibido en payload se considera autoridad.
    - Las colecciones usan paginación por página y orden estable.
    - Fechas y horas se expresan en ISO 8601; la fecha operativa usa la zona
      horaria de la empresa.
    - Importes usan decimal; el servidor recalcula totales.
    - Comandos mobile sensibles requieren idempotencia.
    - Toda operación REST acepta opcionalmente `X-Correlation-Id` conforme a
      `correlation-id-contract.md`, y devuelve el ID efectivo tanto en éxito
      como en error.
    - Auth WEB usa refresh solo en cookie HttpOnly; MOBILE usa body solo fuera
      de contexto navegador. El canal nunca concede identidad, tenant o rol.
  contact:
    name: FollowUpBussiness Engineering
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
  x-contract-status: READY_FOR_HANDOFF
  x-source-of-truth:
    - ../../00_CONTRATO_FUNCIONAL.md
    - ../stories/enablers/EN-013-definir-autenticacion-sesiones-y-recuperacion.md
    - ../architecture/adr/ADR-008-autenticacion-sesiones.md
    - ../stories/BACKLOG_REFINEMENT.md
    - ../stories/dependency-map.md
servers:
  - url: /api/v1
...
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /routes/optimize:
    post:
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      tags: [Routes]
      operationId: optimizeRoute
      summary: Generar propuesta de ruta automática básica
      description: |
        Genera una propuesta versionada y no publicada para una ruta `DRAFT`
        autorizada. Backend deriva fecha, vendedor y territorio de las visitas
        autorizadas; todas deben pertenecer a un único territorio activo asignado
        al vendedor. Los endpoints MVP son la
        ubicación del primer y último `Route.Point` persistidos. El proveedor de matriz recibe solo coordenadas;
        no recibe tenant, vendedor, territorio ni clientes.
      x-required-roles: [COMPANY_ADMIN, SUPERVISOR]
      x-story-ids: [EN-018, BE-022, FE-016, INT-008]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizeRouteRequest' }
      responses:
        '200':
          description: Propuesta optimizada; todavía no publicada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizeRouteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: Problema recuperable con `application/problem+json`, detalle genérico y `correlationId`. Los códigos públicos incluyen `AVAILABILITY_END_MUST_BE_AFTER_START`, `VISIT_WINDOW_END_MUST_BE_AFTER_START`, `VISIT_TERRITORY_REQUIRED`, `VISIT_TERRITORY_INACTIVE`, `VISIT_TERRITORY_NOT_ASSIGNED_TO_SELLER`, `ROUTE_ENDPOINT_LOCATION_REQUIRED` y `MULTIPLE_VISIT_TERRITORIES_NOT_SUPPORTED`; nunca contiene trazas ni detalles internos.
          $ref: '#/components/responses/UnprocessableEntity'
        '503': { $ref: '#/components/responses/RouteOptimizationUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 de acceso, 10 minutos, transportado solo en Authorization; nunca en cookie.
    webRefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-fs-refresh
      description: Cookie refresh exclusiva de WEB; Secure, HttpOnly, SameSite=Strict y Path=/.
    sessionRevocationTicket:
      type: apiKey
      in: header
      name: X-Session-Revocation-Ticket
      description: Ticket opaco MOBILE de un uso y solo para revocar su propia familia; no autentica ni permite refresh.
  headers:
    CorrelationId:
      description: |
        ID efectivo de trazabilidad E2E. Siempre es el UUID canónico definido
        en `correlation-id-contract.md`; es idéntico al campo `correlationId`
        del problema cuando la respuesta es `application/problem+json`.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
        minLength: 36
        maxLength: 36
        example: 018f4e22-7f40-4b9a-a8f9-5e9a3d7b1c20
...
  parameters:
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: |
        ID de trazabilidad opcional aplicable a toda operación REST. El nombre
        de header es insensible a mayúsculas/minúsculas; el valor debe cumplir
        el contrato global. Un valor inválido produce `400 CORRELATION_ID_INVALID`.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
        minLength: 36
        maxLength: 36
...
  responses:
...
    BadRequest:
      description: Solicitud mal formada
      x-error-codes: [CORRELATION_ID_INVALID, VALIDATION_FAILED]
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden:
      description: La identidad no puede realizar la operación o acceder al recurso
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Conflicto de estado, versión o idempotencia
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RouteOptimizationUnavailable:
      description: Motor o proveedor de matriz temporalmente no disponible
      x-error-codes:
        - PROVIDER_UNCONFIGURED
        - PROVIDER_UNAUTHORIZED
        - PROVIDER_RATE_LIMITED
        - PROVIDER_TIMEOUT
        - SOLVER_TIME_LIMIT
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    UnprocessableEntity:
      description: Solicitud sintácticamente válida que incumple reglas de negocio
      headers:
        X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string, maxLength: 200 }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]{2,79}$' }
        detail: { type: string, maxLength: 1000 }
        instance: { type: string, format: uri-reference }
        correlationId: { $ref: '#/components/schemas/CorrelationId' }
        fieldErrors:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/FieldError' }
    CorrelationId:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 018f4e22-7f40-4b9a-a8f9-5e9a3d7b1c20
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string, maxLength: 200 }
        code: { type: string, maxLength: 80 }
        message: { type: string, maxLength: 300 }
...
    OptimizeRouteRequest:
      type: object
      additionalProperties: false
      required:
        - routeId
        - availability
        - baseRouteVersion
        - visits
      properties:
        routeId: { type: string, format: uuid }
        availability: { $ref: '#/components/schemas/TimeWindow' }
        baseRouteVersion:
          type: integer
          minimum: 1
          description: Versión base usada para detectar edición concurrente.
        visits:
          type: array
          minItems: 1
          maxItems: 9
          uniqueItems: true
          description: Los customerId deben ser únicos y pertenecer al mismo tenant y territorio.
          items: { $ref: '#/components/schemas/OptimizeRouteVisitRequest' }
    TimeWindow:
      type: object
      additionalProperties: false
      required: [start, end]
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
          description: Debe ser posterior a start; el backend valida la zona operativa.
    OptimizeRouteVisitRequest:
      type: object
      additionalProperties: false
      required: [customerId, serviceDurationSeconds, priority, windows]
      properties:
        customerId: { type: string, format: uuid }
        serviceDurationSeconds:
          type: integer
          minimum: 1
        priority:
          type: integer
          minimum: 1
          description: Una prioridad mayor aplica mayor penalización por no atender.
        windows:
          type: array
          uniqueItems: true
          description: Ventanas duras; no se relajan silenciosamente. Una lista vacía significa que la visita no tiene restricción horaria.
          items: { $ref: '#/components/schemas/TimeWindow' }
    OptimizeRouteResponse:
      type: object
      additionalProperties: false
      required:
        - routeId
        - proposalVersion
        - baseRouteVersion
        - published
        - orderedVisits
        - unassignedVisits
        - totalTravelSeconds
        - totalServiceSeconds
        - totalDistanceMeters
        - optimality
        - generatedAt
      properties:
        routeId: { type: string, format: uuid }
        proposalVersion: { type: integer, minimum: 1, readOnly: true }
        baseRouteVersion: { type: integer, minimum: 1 }
        published:
          type: boolean
          const: false
          readOnly: true
          description: Optimizar nunca publica una ruta.
        orderedVisits:
          type: array
          items: { $ref: '#/components/schemas/OptimizedRouteVisit' }
        unassignedVisits:
          type: array
          items: { $ref: '#/components/schemas/UnassignedRouteVisit' }
        totalTravelSeconds: { type: integer, minimum: 0 }
        totalServiceSeconds: { type: integer, minimum: 0 }
        totalDistanceMeters: { type: integer, minimum: 0 }
        optimality:
          type: string
          enum: [FEASIBLE, OPTIMAL, TIME_LIMIT]
        generatedAt: { type: string, format: date-time, readOnly: true }
    OptimizedRouteVisit:
      type: object
      additionalProperties: false
      required: [customerId, sequence, plannedArrivalAt, plannedDepartureAt]
      properties:
        customerId: { type: string, format: uuid }
        sequence: { type: integer, minimum: 1 }
        plannedArrivalAt: { type: string, format: date-time }
        plannedDepartureAt: { type: string, format: date-time }
    UnassignedRouteVisit:
      type: object
      additionalProperties: false
      required: [customerId, reason]
      properties:
        customerId: { type: string, format: uuid }
        reason:
          type: string
          enum: [OUTSIDE_SHIFT, TIME_WINDOW_CONFLICT, UNREACHABLE, LIMIT_EXCEEDED]
...